Move Node into its own module.
